+++ Assets/Project/Modules/CombatSystem/Scripts/DamageHit/DamageHitConfig.cs
@@ -13,7 +13,7 @@ namespace Project.Modules.CombatSystem
         
         
         public DamageHitTargetType DamageHitTargetTypeMask => _damageHitPreset.TargetMask;
-        public float Damage => _damage;
+        public int Damage => _damage;
         public float KnockbackMagnitude => _knockbackMagnitude;
         public float StunDuration => _stunDuration;
